idp/src/main/java/eu/scape_project/pw/idp/validator/ConfirmPasswordValidator.java
/*******************************************************************************
 * Copyright 2006 - 2012 Vienna University of Technology,
 * Department of Software Technology and Interactive Systems, IFS
 * 
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 *   http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ******************************************************************************/
package eu.scape_project.pw.idp.validator;

import javax.faces.application.FacesMessage;
import javax.faces.component.UIComponent;
import javax.faces.context.FacesContext;
import javax.faces.validator.FacesValidator;
import javax.faces.validator.Validator;
import javax.faces.validator.ValidatorException;

@FacesValidator("ConfirmPasswordValidator")
public class ConfirmPasswordValidator implements Validator {

  @Override
  public void validate(FacesContext context, UIComponent component, Object value) throws ValidatorException {
    String password = (String) value;
    String confirm = (String) component.getAttributes().get("pwConfirm");

    // Just ignore and let required="true" do its job.
    if (password == null || confirm == null || password.length() == 0 || confirm.length() == 0) {
      return;
    }

    if (!password.equals(confirm)) {
      throw new ValidatorException(new FacesMessage("Passwords are not equal."));
    }
  }
}
